% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity.R
\name{annuity}
\alias{annuity}
\title{Present and Future Values of an Annuity}
\usage{
annuity(pmt, int_rate, growth, n)
}
\arguments{
\item{pmt}{payment}

\item{int_rate}{interest rate}

\item{growth}{growth rate}

\item{n}{number of years}
}
\description{
This function calculates annuities
}
\examples{
annuity()
}
\keyword{annuity}

