% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_price.R
\name{bond_price}
\alias{bond_price}
\title{Price of different kinds of Bonds}
\usage{
bond_price(C, r, n, Fv)
}
\arguments{
\item{C}{coupon}

\item{r}{rate}

\item{n}{number of years}

\item{Fv}{face value}
}
\description{
This function calculates bond prices
}
\examples{
bond_price()
}
\keyword{bond}
\keyword{consol}
\keyword{coupon}
\keyword{discount}

