% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_int.R
\name{effective_int}
\alias{effective_int}
\title{Effective Interest Rate}
\usage{
effective_int(int_rate, m)
}
\arguments{
\item{int_rate}{interest rate}

\item{m}{times compunding per year, "inf" if continuous}
}
\description{
This function calculates effective interest rates
}
\examples{
effective_int()
}
\keyword{effective}
\keyword{interest}

