% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_val.R
\name{future_val}
\alias{future_val}
\title{Future Value Calculator}
\usage{
future_val(present_val, int_rate, m, n)
}
\arguments{
\item{int_rate}{interest rate}

\item{m}{times compounding per year, "inf" if continuous}

\item{n}{number of years}

\item{present_value}{present value at time 0}
}
\description{
This function calculates future values
}
\examples{
future_val()
}
\keyword{future}
\keyword{value}

