% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_returns.R
\name{mean_returns}
\alias{mean_returns}
\title{Return Means (Arithmetic and Geometric)}
\usage{
mean_returns(t, r)
}
\arguments{
\item{t}{number of periods}

\item{r}{vector of returns}
}
\description{
This function calculates return means
}
\examples{
mean_returns()
}
\keyword{arithmetic}
\keyword{geometric}
\keyword{mean}
\keyword{returns}

