% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom2real.R
\name{nom2real}
\alias{nom2real}
\title{Real Return Function}
\usage{
nom2real(nom_rate, inf_rate)
}
\arguments{
\item{nom_rate}{nominal rate}

\item{inf_rate}{inflation rate}
}
\description{
This function calculates real returns given nominal
}
\examples{
nom2real()
}
\keyword{inflation}
\keyword{nominal}

