% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetuity.R
\name{perpetuity}
\alias{perpetuity}
\title{Present Value of a Perpetuity}
\usage{
perpetuity(pmt, int_rate, growth)
}
\arguments{
\item{pmt}{payment}

\item{int_rate}{interest rate}

\item{growth}{growth rate}
}
\description{
This function calculates perpetuities
}
\examples{
perpetuity()
}
\keyword{perpetuity}

