% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_weights_returns.R
\name{portfolio_weights_returns}
\alias{portfolio_weights_returns}
\title{Portfolio Weights - 2 risky assets and 1 risk-free asset}
\usage{
portfolio_weights_returns(var_x, var_y, cov_x_y, rf, return_x, return_y,
  weight_x, gamma)
}
\arguments{
\item{var_x}{variance of x}

\item{var_y}{variance of y}

\item{cov_x_y}{covariance of x and y}

\item{rf}{risk-free rate}

\item{return_x}{mean return of x}

\item{return_y}{mean return of y}

\item{weight_x}{weight on x if one risky and one risk free}

\item{gamma}{risk aversion coefficient}
}
\description{
This function calculates portfolio weights for two risky assets and a risk free asset
}
\examples{
portfolio_weights_returns()
}
\keyword{mean}
\keyword{risk}
\keyword{tangency}
\keyword{variance}

