% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_val.R
\name{present_val}
\alias{present_val}
\title{Present Value Calculator}
\usage{
present_val(future_val, int_rate, m, n)
}
\arguments{
\item{int_rate}{interest rate}

\item{m}{times compunding per year, "inf" if continuous}

\item{n}{number of years}

\item{future_value}{future value at time 0}
}
\description{
This function calculates present values
}
\examples{
present_val()
}
\keyword{present}
\keyword{value}

